nft-tda-reanalysis/requirements.txt
mne>=1.6
giotto-tda>=0.6
numpy>=1.24
scipy>=1.11
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
openneuro-py>=2024.1
joblib>=1.3
statsmodels>=0.14
